# src/utility/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Social Network for Music API",
    "description": "API per il sito Social Network for Music",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/search-tracks": {
      "get": {
        "tags": [
          "Spotify"
        ],
        "summary": "Ricerca di una traccia",
        "description": "",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/get-track/{id}": {
      "get": {
        "tags": [
          "Spotify"
        ],
        "summary": "Restituisce le informazioni su una traccia",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/search-artist": {
      "get": {
        "tags": [
          "Spotify"
        ],
        "summary": "Ricerca di un artista",
        "description": "",
        "parameters": [
          {
            "name": "artist",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/get-artist-tracks": {
      "get": {
        "tags": [
          "Spotify"
        ],
        "summary": "Restituisce le tracce di un artista",
        "description": "",
        "parameters": [
          {
            "name": "artistId",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/get-available-genres": {
      "get": {
        "tags": [
          "Spotify"
        ],
        "summary": "Restituisce i generi disponibili su Spotify",
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/recommendations": {
      "get": {
        "tags": [
          "Spotify"
        ],
        "summary": "Restituisce le raccomandazioni per l'utente in base ai suoi generi e artisti preferiti",
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/recommendations-genre": {
      "get": {
        "tags": [
          "Spotify"
        ],
        "summary": "Restituisce le raccomandazioni per l'utente in base al genere fornito",
        "description": "",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  }
}